move JSendMarshallingSupport to webresult, rework dependencies
